move things to make one-ring-dist independent of one-ring-cli, also bump version to 2.2.0 due to all that kerfuffle
