+++ Commons/src/main/java/ash/nazg/config/WrapperConfigBuilder.java
@@ -2,7 +2,7 @@
  * Copyright (C) 2020 Locomizer team and Contributors
  * This project uses New BSD license with do no evil clause. For full text, check the LICENSE file in the root directory.
  */
-package ash.nazg.cli.config;
+package ash.nazg.config;
 
 import ash.nazg.config.InvalidConfigValueException;
 import ash.nazg.config.PropertiesConfig;
