+++ DateTime/pom.xml
@@ -6,17 +6,17 @@
     <parent>
         <groupId>ash.nazg</groupId>
         <artifactId>OneRing</artifactId>
-        <version>2.1.0</version>
+        <version>2.2.0</version>
     </parent>
 
     <artifactId>DateTime</artifactId>
-    <version>2.1.0</version>
+    <version>2.2.0</version>
 
     <dependencies>
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Commons</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
 
         <dependency>
@@ -28,7 +28,7 @@
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Commons</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
