+++ DistWrapper/pom.xml
@@ -7,11 +7,11 @@
     <parent>
         <artifactId>OneRing</artifactId>
         <groupId>ash.nazg</groupId>
-        <version>2.1.0</version>
+        <version>2.2.0</version>
     </parent>
 
     <artifactId>DistWrapper</artifactId>
-    <version>2.1.0</version>
+    <version>2.2.0</version>
 
     <build>
         <plugins>
@@ -55,7 +55,8 @@
                     <properties>
                         <license.years>2020</license.years>
                         <license.holder>Locomizer team and Contributors</license.holder>
-                        <license.ref>This project uses New BSD license with do no evil clause. For full text, check the LICENSE file in the root directory.
+                        <license.ref>This project uses New BSD license with do no evil clause. For full text, check the
+                            LICENSE file in the root directory.
                         </license.ref>
                     </properties>
                     <includes>
@@ -70,8 +71,8 @@
     <dependencies>
         <dependency>
             <groupId>ash.nazg</groupId>
-            <artifactId>TaskWrapper</artifactId>
-            <version>2.1.0</version>
+            <artifactId>Commons</artifactId>
+            <version>2.2.0</version>
         </dependency>
         <dependency>
             <groupId>org.apache.parquet</groupId>
@@ -82,7 +83,7 @@
 
     <profiles>
         <profile>
-            <id>default</id>
+            <id>emr</id>
             <activation>
                 <activeByDefault>true</activeByDefault>
             </activation>
@@ -102,9 +103,31 @@
                                     <transformers>
                                         <transformer
                                                 implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
-                                            <mainClass>ash.nazg.dist.Dist</mainClass>
+                                            <mainClass>ash.nazg.dist.Main</mainClass>
                                         </transformer>
                                     </transformers>
+                                    <artifactSet>
+                                        <excludes>
+                                            <exclude>org.apache.hadoop</exclude>
+                                            <exclude>commons-collections</exclude>
+                                            <exclude>commons-logging</exclude>
+                                            <exclude>org.apache.commons:commons-lang3</exclude>
+                                            <exclude>org.apache.spark</exclude>
+                                            <exclude>org.scala-lang</exclude>
+                                        </excludes>
+                                        <includes>
+                                            <include>ash.nazg</include>
+                                            <include>com.opencsv</include>
+                                            <include>org.locationtech.jts</include>
+                                            <include>org.wololo</include>
+                                            <include>io.github.classgraph</include>
+                                            <include>com.github.alexmojaki</include>
+                                            <include>net.sf.geographiclib</include>
+                                            <include>commons-cli</include>
+                                            <include>com.uber</include>
+                                            <include>org.apache.parquet</include>
+                                        </includes>
+                                    </artifactSet>
                                 </configuration>
                             </execution>
                         </executions>
@@ -118,6 +141,31 @@
             <activation>
                 <activeByDefault>false</activeByDefault>
             </activation>
+            <build>
+                <plugins>
+                    <plugin>
+                        <artifactId>maven-shade-plugin</artifactId>
+                        <version>3.2.1</version>
+                        <executions>
+                            <execution>
+                                <phase>package</phase>
+                                <goals>
+                                    <goal>shade</goal>
+                                </goals>
+                                <configuration>
+                                    <createDependencyReducedPom>false</createDependencyReducedPom>
+                                    <transformers>
+                                        <transformer
+                                                implementation="org.apache.maven.plugins.shade.resource.ManifestResourceTransformer">
+                                            <mainClass>ash.nazg.dist.Main</mainClass>
+                                        </transformer>
+                                    </transformers>
+                                </configuration>
+                            </execution>
+                        </executions>
+                    </plugin>
+                </plugins>
+            </build>
         </profile>
     </profiles>
 </project>
