+++ Geohashing/pom.xml
@@ -7,17 +7,17 @@
     <parent>
         <artifactId>OneRing</artifactId>
         <groupId>ash.nazg</groupId>
-        <version>2.1.0</version>
+        <version>2.2.0</version>
     </parent>
 
     <artifactId>Geohashing</artifactId>
-    <version>2.1.0</version>
+    <version>2.2.0</version>
 
     <dependencies>
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Commons</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
 
         <dependency>
@@ -29,7 +29,7 @@
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Commons</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
