+++ TaskWrapper/pom.xml
@@ -6,48 +6,48 @@
     <parent>
         <groupId>ash.nazg</groupId>
         <artifactId>OneRing</artifactId>
-        <version>2.1.0</version>
+        <version>2.2.0</version>
     </parent>
 
     <artifactId>TaskWrapper</artifactId>
-    <version>2.1.0</version>
+    <version>2.2.0</version>
 
     <dependencies>
         <!-- One Ring Spark modules. Add all new modules here -->
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Columnar</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>SimpleFilters</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Proximity</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>DateTime</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Geohashing</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Math</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Populations</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
         </dependency>
         <!-- end -->
 
@@ -101,7 +101,7 @@
         <dependency>
             <groupId>ash.nazg</groupId>
             <artifactId>Commons</artifactId>
-            <version>2.1.0</version>
+            <version>2.2.0</version>
             <type>test-jar</type>
             <scope>test</scope>
         </dependency>
@@ -213,7 +213,7 @@
 
     <profiles>
         <profile>
-            <id>default</id>
+            <id>emr</id>
             <activation>
                 <activeByDefault>true</activeByDefault>
             </activation>
