+++ pom.xml
@@ -4,7 +4,7 @@
 
     <groupId>ash.nazg</groupId>
     <artifactId>OneRing</artifactId>
-    <version>2.1.0</version>
+    <version>2.2.0</version>
     <name>One Ring Spark Modules root</name>
     <url>http://github.com/PastorGL/one-ring-oss</url>
     <packaging>pom</packaging>
